Se agregaron las siguientes funcionalidades, Se aplicó la vista parcial de no resultados en las vistas asociadas al estudiante.
Correciónes de bugs:
-Bug que detenía la aplicación al terminar un desafío con id de proyecto no válido.
-Mejoras de rendimiento en algunos controladores.
-Mejoras de código en la vista parcial de no resultados.

+++ Hera/Data/DataAccess_Sql.cs
@@ -115,7 +115,7 @@ namespace Hera.Data
         public async Task Delete_Desafio(int id)
         {
             var desafio = await Find_Desafio(id);
-            if(desafio != null)
+            if(desafio != null && desafio.Popularity == 0)
             {
                 _fmService.DeleteFile(desafio.DirSolucion);
                 Delete<Desafio>(desafio);
